refactor(module-source): Converge file naming convention
