+++ packages/module-source/src/babel-plugin.js
